pyproject.toml
[project]
name = "tuim"
version = "0.0.1-beta"
description = "TUI Remote Connection Manager - k9s-style terminal UI for SSH, RDP, VNC, Telnet and K8s"
requires-python = ">=3.14"
dependencies = [
    "textual>=0.80.0",
    "pyyaml>=6.0",
    "asyncssh>=2.14.0",
    "telnetlib3>=2.0",
    "pyte>=0.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "textual-dev>=1.0"]

[project.scripts]
tuim = "tuim.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tuim"]
